Separate fileManager and guFile, do everything through fileManager
